requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
diff-match-patch==20241021
distro==1.9.0
Django==5.2.8
django-debug-toolbar==4.2.0
django-import-export==4.3.14
django-widget-tweaks==1.5.0
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0  # ADD THIS LINE - essential for deployment
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
openai==2.8.1
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.1.0
python-docx==1.2.0
python-dotenv==1.2.1
reportlab==4.4.5
sniffio==1.3.1
sqlparse==0.5.4
tablib==3.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
